+++ sdk/servicebus/service-bus/samples/javascript/advanced/processMessageFromDLQ.js
@@ -33,7 +33,7 @@ async function main() {
 
 async function processDeadletterMessageQueue() {
   // If connecting to a subscription's dead letter queue you can use the createDeadLetterReceiver(topic, subscription) overload
-  const receiver = sbClient.createDeadLetterReceiver(queueName, "peekLock");
+  const receiver = sbClient.createDeadLetterReceiver(queueName);
 
   const messages = await receiver.receiveMessages(1);