+++ sdk/servicebus/service-bus/src/receivers/sessionReceiver.ts
@@ -128,7 +128,9 @@ export class SessionReceiverImpl<ReceivedMessageT extends ReceivedMessage | Rece
   >(
     context: ClientEntityContext,
     receiveMode: "peekLock" | "receiveAndDelete",
-    sessionOptions: CreateSessionReceiverOptions,
+    sessionOptions:
+      | CreateSessionReceiverOptions<"peekLock">
+      | CreateSessionReceiverOptions<"receiveAndDelete">,
     retryOptions: RetryOptions = {}
   ): Promise<SessionReceiver<ReceivedMessageT>> {
     context.isSessionEnabled = true;