+++ sdk/servicebus/service-bus/src/serviceBusAtomManagementClient.ts
@@ -2332,7 +2332,7 @@ export class ServiceBusManagementClient extends ServiceClient {
         const err = new RestError(
           `The messaging entity "${name}" being requested cannot be found.`,
           "MessageEntityNotFoundError",
-          404,
+          response.status,
           stripRequest(webResource),
           stripResponse(response)
         );