+++ sdk/servicebus/service-bus/test/serviceBusClient.spec.ts
@@ -62,12 +62,8 @@ describe("Random scheme in the endpoint from connection string", function(): voi
     );
     const sender = sbClientWithRelaxedEndPoint.createSender(entities.queue || entities.topic!);
     const receiver = entities.queue
-      ? sbClientWithRelaxedEndPoint.createReceiver(entities.queue, "peekLock")
-      : sbClientWithRelaxedEndPoint.createReceiver(
-          entities.topic!,
-          entities.subscription!,
-          "peekLock"
-        );
+      ? sbClientWithRelaxedEndPoint.createReceiver(entities.queue)
+      : sbClientWithRelaxedEndPoint.createReceiver(entities.topic!, entities.subscription!);
 
     // Send and receive messages
     const testMessages = entities.usesSessions
@@ -133,7 +129,7 @@ describe("Errors with non existing Namespace", function(): void {
   it("throws error when receiving batch data to a non existing namespace", async function(): Promise<
     void
   > {
-    const receiver = sbClient.createReceiver("some-queue", "peekLock");
+    const receiver = sbClient.createReceiver("some-queue");
     await receiver.receiveMessages(10).catch(testError);
 
     should.equal(errorWasThrown, true, "Error thrown flag must be true");
@@ -142,7 +138,7 @@ describe("Errors with non existing Namespace", function(): void {
   it("throws error when receiving streaming data from a non existing namespace", async function(): Promise<
     void
   > {
-    const receiver = sbClient.createReceiver("some-queue", "peekLock");
+    const receiver = sbClient.createReceiver("some-queue");
     receiver.subscribe({
       async processMessage() {
         throw "processMessage should not have been called when receive call is made from a non existing namespace";
@@ -195,7 +191,7 @@ describe("Errors with non existing Queue/Topic/Subscription", async function(): 
   it("throws error when receiving batch data from a non existing queue", async function(): Promise<
     void
   > {
-    const receiver = sbClient.createReceiver("some-name", "peekLock");
+    const receiver = sbClient.createReceiver("some-name");
     await receiver.receiveMessages(1).catch((err) => testError(err, "some-name"));
 
     should.equal(errorWasThrown, true, "Error thrown flag must be true");
@@ -204,11 +200,7 @@ describe("Errors with non existing Queue/Topic/Subscription", async function(): 
   it("throws error when receiving batch data from a non existing subscription", async function(): Promise<
     void
   > {
-    const receiver = sbClient.createReceiver(
-      "some-topic-name",
-      "some-subscription-name",
-      "peekLock"
-    );
+    const receiver = sbClient.createReceiver("some-topic-name", "some-subscription-name");
     await receiver
       .receiveMessages(1)
       .catch((err) => testError(err, "some-topic-name/Subscriptions/some-subscription-name"));
@@ -219,7 +211,7 @@ describe("Errors with non existing Queue/Topic/Subscription", async function(): 
   it("throws error when receiving streaming data from a non existing queue", async function(): Promise<
     void
   > {
-    const receiver = sbClient.createReceiver("some-name", "peekLock");
+    const receiver = sbClient.createReceiver("some-name");
     receiver.subscribe({
       async processMessage() {
         throw "processMessage should not have been called when receive call is made from a non existing namespace";
@@ -240,11 +232,7 @@ describe("Errors with non existing Queue/Topic/Subscription", async function(): 
   it("throws error when receiving streaming data from a non existing subscription", async function(): Promise<
     void
   > {
-    const receiver = sbClient.createReceiver(
-      "some-topic-name",
-      "some-subscription-name",
-      "peekLock"
-    );
+    const receiver = sbClient.createReceiver("some-topic-name", "some-subscription-name");
     receiver.subscribe({
       async processMessage() {
         throw "processMessage should not have been called when receive call is made from a non existing namespace";
@@ -351,8 +339,8 @@ describe("Test ServiceBusClient with TokenCredentials", function(): void {
         const sbClient = new ServiceBusClient(serviceBusEndpoint, tokenCreds);
         const sender = sbClient.createSender(entities.queue || entities.topic!);
         const receiver = entities.queue
-          ? sbClient.createReceiver(entities.queue!, "peekLock")
-          : sbClient.createReceiver(entities.topic!, entities.subscription!, "peekLock");
+          ? sbClient.createReceiver(entities.queue!)
+          : sbClient.createReceiver(entities.topic!, entities.subscription!);
 
         const testMessages = TestMessage.getSample();
         await sender.sendMessages(testMessages);
@@ -712,7 +700,7 @@ describe("entityPath on sender and receiver", async () => {
 
   it("Entity Path on Queue Receiver", () => {
     const dummyQueueName = "dummy";
-    const receiver = sbClient.createReceiver(dummyQueueName, "peekLock");
+    const receiver = sbClient.createReceiver(dummyQueueName);
     should.equal(
       receiver.entityPath,
       dummyQueueName,
@@ -722,7 +710,7 @@ describe("entityPath on sender and receiver", async () => {
 
   it("Entity Path on Queue deadletter Receiver", () => {
     const dummyQueueName = "dummy";
-    const receiver = sbClient.createDeadLetterReceiver(dummyQueueName, "peekLock");
+    const receiver = sbClient.createDeadLetterReceiver(dummyQueueName);
     should.equal(
       receiver.entityPath,
       `${dummyQueueName}/$DeadLetterQueue`,
@@ -733,7 +721,7 @@ describe("entityPath on sender and receiver", async () => {
   it("Entity Path on Subscription Receiver", () => {
     const dummyTopicName = "dummyTopicName";
     const dummySubscriptionName = "dummySubscriptionName";
-    const receiver = sbClient.createReceiver(dummyTopicName, dummySubscriptionName, "peekLock");
+    const receiver = sbClient.createReceiver(dummyTopicName, dummySubscriptionName);
     should.equal(
       receiver.entityPath,
       `${dummyTopicName}/Subscriptions/${dummySubscriptionName}`,
@@ -744,11 +732,7 @@ describe("entityPath on sender and receiver", async () => {
   it("Entity Path on Subscription deadletter Receiver", () => {
     const dummyTopicName = "dummyTopicName";
     const dummySubscriptionName = "dummySubscriptionName";
-    const receiver = sbClient.createDeadLetterReceiver(
-      dummyTopicName,
-      dummySubscriptionName,
-      "peekLock"
-    );
+    const receiver = sbClient.createDeadLetterReceiver(dummyTopicName, dummySubscriptionName);
     should.equal(
       receiver.entityPath,
       `${dummyTopicName}/Subscriptions/${dummySubscriptionName}/$DeadLetterQueue`,